Refactor metrics queries to use local repo.
