+++ apps/explorer/lib/explorer/chain.ex
@@ -2172,46 +2172,7 @@ defmodule Explorer.Chain do
       0
   end
 
-  @doc """
-  Fetches count of last n blocks, last block timestamp, last block number and average gas used in the last minute.
-  Using a single method to fetch and calculate these values for performance reasons (only 2 queries used).
-  """
-  @spec metrics_fetcher(integer | nil) ::
-          {non_neg_integer, non_neg_integer, non_neg_integer, float}
-  def metrics_fetcher(n) do
-    last_block_number = fetch_max_block_number()
 
-    if last_block_number == 0 do
-      {0, 0, 0, 0}
-    else
-      range_start = last_block_number - n + 1
-
-      last_n_blocks_result =
-        SQL.query!(
-          Repo,
-          """
-          SELECT
-          COUNT(*) AS last_n_blocks_count,
-          CAST(EXTRACT(EPOCH FROM (DATE_TRUNC('second', NOW()::timestamp) - MAX(timestamp))) AS INTEGER) AS last_block_age,
-          AVG((gas_used/gas_limit)*100) AS average_gas_used
-          FROM blocks
-          WHERE number BETWEEN $1 AND $2;
-          """,
-          [range_start, last_block_number]
-        )
-
-      {last_n_blocks_count, last_block_age, average_gas_used} =
-        case Map.fetch(last_n_blocks_result, :rows) do
-          {:ok, [[last_n_blocks_count, last_block_age, average_gas_used]]} ->
-            {last_n_blocks_count, last_block_age, average_gas_used}
-
-          _ ->
-            0
-        end
-
-      {last_n_blocks_count, last_block_age, last_block_number, Decimal.to_float(average_gas_used)}
-    end
-  end
 
   @spec fetch_count_consensus_block() :: non_neg_integer
   def fetch_count_consensus_block do
@@ -7357,77 +7318,7 @@ defmodule Explorer.Chain do
 
   defp boolean_to_check_result(false), do: :not_found
 
-  @spec fetch_number_of_locks :: non_neg_integer()
-  def fetch_number_of_locks do
-    result =
-      SQL.query(Repo, """
-      SELECT COUNT(*) FROM (SELECT blocked_locks.pid     AS blocked_pid,
-           blocked_activity.usename  AS blocked_user,
-           blocking_locks.pid     AS blocking_pid,
-           blocking_activity.usename AS blocking_user,
-           blocked_activity.query    AS blocked_statement,
-           blocking_activity.query   AS current_statement_in_blocking_process
-      FROM  pg_catalog.pg_locks         blocked_locks
-      JOIN pg_catalog.pg_stat_activity blocked_activity  ON blocked_activity.pid = blocked_locks.pid
-      JOIN pg_catalog.pg_locks         blocking_locks
-          ON blocking_locks.locktype = blocked_locks.locktype
-          AND blocking_locks.DATABASE IS NOT DISTINCT FROM blocked_locks.DATABASE
-          AND blocking_locks.relation IS NOT DISTINCT FROM blocked_locks.relation
-          AND blocking_locks.page IS NOT DISTINCT FROM blocked_locks.page
-          AND blocking_locks.tuple IS NOT DISTINCT FROM blocked_locks.tuple
-          AND blocking_locks.virtualxid IS NOT DISTINCT FROM blocked_locks.virtualxid
-          AND blocking_locks.transactionid IS NOT DISTINCT FROM blocked_locks.transactionid
-          AND blocking_locks.classid IS NOT DISTINCT FROM blocked_locks.classid
-          AND blocking_locks.objid IS NOT DISTINCT FROM blocked_locks.objid
-          AND blocking_locks.objsubid IS NOT DISTINCT FROM blocked_locks.objsubid
-          AND blocking_locks.pid != blocked_locks.pid
-      JOIN pg_catalog.pg_stat_activity blocking_activity ON blocking_activity.pid = blocking_locks.pid
-      WHERE NOT blocked_locks.GRANTED) a;
-      """)
 
-    case result do
-      {:ok, %Postgrex.Result{rows: [[rows]]}} -> rows
-      _ -> 0
-    end
-  end
-
-  @spec fetch_number_of_dead_locks :: non_neg_integer()
-  def fetch_number_of_dead_locks do
-    database =
-      :explorer
-      |> Application.get_env(Explorer.Repo)
-      |> Keyword.get(:database)
-
-    result =
-      SQL.query(
-        Repo,
-        """
-        SELECT deadlocks FROM pg_stat_database where datname = $1;
-        """,
-        [database]
-      )
-
-    case result do
-      {:ok, %Postgrex.Result{rows: [[rows]]}} -> rows
-      _ -> 0
-    end
-  end
-
-  @spec fetch_name_and_duration_of_longest_query :: non_neg_integer()
-  def fetch_name_and_duration_of_longest_query do
-    result =
-      SQL.query(Repo, """
-        SELECT query, NOW() - xact_start AS duration FROM pg_stat_activity
-        WHERE state IN ('idle in transaction', 'active') ORDER BY now() - xact_start DESC LIMIT 1;
-      """)
-
-    {:ok, longest_query_map} = result
-
-    case Map.fetch(longest_query_map, :rows) do
-      {:ok, [[_, longest_query_duration]]} when not is_nil(longest_query_duration) -> longest_query_duration.secs
-      _ -> 0
-    end
-  end
 
   @doc """
   Fetches the first trace from the Parity trace URL.
